reorganize deployment scripts
