+++ deploy/arbitrum/003_deploy_AmplificationUtils.ts
@@ -7,14 +7,6 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
   const { deploy } = deployments
   const { libraryDeployer } = await getNamedAccounts()
 
-  if (isTestNetwork(await getChainId())) {
-    await deploy("AmplificationUtilsV1", {
-      from: libraryDeployer,
-      log: true,
-      skipIfAlreadyDeployed: true,
-    })
-  }
-
   await deploy("AmplificationUtils", {
     from: libraryDeployer,
     log: true,
