+++ deploy/mainnet/000_prepare_network.ts
@@ -1,6 +1,6 @@
 import { HardhatRuntimeEnvironment } from "hardhat/types"
 import { DeployFunction } from "hardhat-deploy/types"
-import { isMainnet } from "../utils/network"
+import { isMainnet } from "../../utils/network"
 import dotenv from "dotenv"
 
 const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
