+++ deploy/mainnet/140_check_alETHPoolTokens.ts
@@ -1,6 +1,6 @@
 import { HardhatRuntimeEnvironment } from "hardhat/types"
 import { DeployFunction } from "hardhat-deploy/types"
-import { isTestNetwork } from "../utils/network"
+import { isTestNetwork } from "../../utils/network"
 import { BigNumber } from "ethers"
 
 const ALETH_TOKENS_ARGS: { [token: string]: any[] } = {
