+++ deploy/mainnet/170_deploy_SwapMigrator.ts
@@ -1,7 +1,7 @@
 import { HardhatRuntimeEnvironment } from "hardhat/types"
 import { DeployFunction } from "hardhat-deploy/types"
-import { MULTISIG_ADDRESS } from "../utils/accounts"
-import { isTestNetwork } from "../utils/network"
+import { MULTISIG_ADDRESS } from "../../utils/accounts"
+import { isTestNetwork } from "../../utils/network"
 
 const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
   const { deployments, getNamedAccounts, getChainId } = hre
