+++ deploy/mainnet/200_check_TBTCMetaPoolTokens.ts
@@ -1,6 +1,6 @@
 import { DeployFunction } from "hardhat-deploy/types"
 import { HardhatRuntimeEnvironment } from "hardhat/types"
-import { isTestNetwork } from "../utils/network"
+import { isTestNetwork } from "../../utils/network"
 import { BigNumber } from "ethers"
 
 const USD_TOKENS_ARGS: { [token: string]: any[] } = {
