+++ deploy/mainnet/210_check_CUSDMetaPoolTokens.ts
@@ -1,7 +1,7 @@
 import { BigNumber } from "ethers"
 import { DeployFunction } from "hardhat-deploy/types"
 import { HardhatRuntimeEnvironment } from "hardhat/types"
-import { isTestNetwork } from "../utils/network"
+import { isTestNetwork } from "../../utils/network"
 
 const USD_TOKENS_ARGS: { [token: string]: any[] } = {
   WCUSD: ["Wrapped Celo USD", "wCUSD", "18"],
